Confirmed that airquality class works on apiApp

+++ src/main/java/cs1302/api/AirQuality.java
@@ -25,7 +25,7 @@ public class AirQuality {
         int tp;
         int pr;
         int hu;
-        int ws;
+        double ws;
         int wd;
         String ic;
     }
